Creates Login component - Creates new Login component - Refactors App and moves login-functionality to the Login component
