+++ cypress/fixtures/unsuccessful_user_login.json
@@ -1,3 +1,3 @@
 {
-  "message": "Invalid login credentials. Please try again"
+  "errors": "Invalid login credentials. Please try again"
 }
